refactor: Move MemoryRecorder to its own module.
